requirements.txt
Django==5.0.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.1.0
djangorestframework-simplejwt==5.3.1
isort==5.13.2
django-debug-toolbar==4.2.0
flake8==7.0.0
black==24.1.1